+++ daeng/src/routes/AppRouter.jsx
@@ -11,8 +11,9 @@ import MyReviewPage from '../pages/review/MyReviewPage';
 import PlaceDetail from '../pages/PlaceDetail';
 import TotalReviewPage from '../pages/review/TotalReviewPage';
 import WriteReviewPage from '../pages/review/WriteReviewPage';
-import Preference from '../pages/Preference';
+import Preference from '../pages/user/Preference';
 import ScrollTop from '../components/commons/ScrollTop';
+import EditPreference from '../pages/user/EditPreference';
 
 const AppRouter = () => {
   return (
@@ -21,7 +22,8 @@ const AppRouter = () => {
       <Routes>
         <Route path="/" element={<Home />} />
         <Route path="/login" element={<Login />} />
-        <Route path="/preference" element={<Preference />} />
+        <Route path="/preference-register" element={<Preference />} />
+        <Route path="/preference-edit" element={<EditPreference />} />
         <Route path="/search" element={<Search />} />
         <Route path="/search/:id" element={<PlaceDetail />} />
         <Route path="/pet-register" element={<PetRegisterPage />} />
